:recycle: update option amount from uint80 to uint64

+++ src/interface/option/IOptionActions.sol
@@ -7,19 +7,19 @@ interface IOptionActions {
         address quoteAsset,
         uint32[] calldata exerciseWindow,
         uint256 strikePrice,
-        uint80 optionAmount
+        uint64 optionAmount
     ) external returns (uint256 optionTokenId);
     function writePut(
         address baseAsset,
         address quoteAsset,
         uint32[] calldata exerciseWindow,
         uint256 strikePrice,
-        uint80 optionAmount
+        uint64 optionAmount
     ) external returns (uint256 optionTokenId);
-    function write(uint256 optionTokenId, uint80 optionAmount) external;
-    function batchWrite(uint256[] calldata optionTokenIds, uint80[] calldata optionAmounts) external;
-    function exercise(uint256 optionTokenId, uint80 optionAmount) external;
-    function netOff(uint256 optionTokenId, uint80 optionAmount)
+    function write(uint256 optionTokenId, uint64 optionAmount) external;
+    function batchWrite(uint256[] calldata optionTokenIds, uint64[] calldata optionAmounts) external;
+    function exercise(uint256 optionTokenId, uint64 optionAmount) external;
+    function netOff(uint256 optionTokenId, uint64 optionAmount)
         external
         returns (uint256 writeAssetNettedOff); // TODO width needs to change
     function redeem(uint256 optionTokenId)
